show an error info when trying to delete a category that has news

+++ src/main/java/io/github/hrashk/news/api/categories/CategoryService.java
@@ -1,5 +1,7 @@
 package io.github.hrashk.news.api.categories;
 
+import io.github.hrashk.news.api.news.NewsRepository;
+import jakarta.validation.ValidationException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -10,6 +12,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CategoryService {
     private final CategoryRepository repository;
+    private final NewsRepository newsRepository;
 
     public List<Category> findAll(Pageable pageable) {
         return repository.findAll(pageable).getContent();
@@ -24,6 +27,9 @@ public class CategoryService {
     }
 
     public void delete(Category category) {
+        if (newsRepository.existsByCategory(category))
+            throw new ValidationException("Cannot delete the category because it has related news");
+
         repository.delete(category);
     }
 }
